requirements.txt
# Requirements for Raspberry Pi Speed Monitor (SMB Version)
# Install in virtual environment to avoid "Externally-managed-environment" error
# 
# Setup:
#   python3 -m venv ~/.venv/speedtest_monitor
#   source ~/.venv/speedtest_monitor/bin/activate
#   pip install -r requirements.txt

# Core dependencies
speedtest-cli>=2.1.3

# Web UI dependencies
flask>=2.3.0
flask-cors>=4.0.0

# Optional: For data analysis and visualization
pandas>=1.3.0
numpy>=1.21.0

# Optional: For email notifications
secure-smtplib>=0.1.1

# Note: No cloud dependencies needed with SMB approach
# Data is synced to local network share via file copying

# Development dependencies (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0